requirements.txt
openai'
google.generativeai
transformers